Merge PR #28: Produce ibctest binary built with go test
* Begin relayertest package

This package exports a TestRelayer function that is both used internally
in ibc-test-framework and may be imported by other packages to test an
implementation of a relayer.

* Port IBCTestCase.RelayPacketTest to Go unit test

And update cmd/custom.go and cmd/test.go to support calling those
functions directly.

* Remove no-longer-valid test

* Add cmd/ibctest to provide a test binary

There would be a bit of remaining work to get this production-ready.

- Remove the existing cmd package?
- Port the remaining ibc.IBCTest tests to the relayertest package
- Add a Makefile to build the ibctest binary
- Fix the moniker-too-long error instead of working around it
- Support custom chain factories instead of just the builtins

* Remove existing cmd package

We are now headed in the direction of only providing a go test-built
binary, so remove a set of code that we no longer need.

* Make cosmos node monikers fit in character limit

This allows effectively unlimited subtest structures now.

The fitting algorithm was arbitrarily chosen to keep the start and end
of the test name, and append a hash to the end. It is possible that we
may want a different algorithm at some other point in time, that splits
on / and keeps several segments, so that a reader gets a better sense of
the subtest tree. For now, this approach should suffice.

* Improve sanitizeTestNameForContainer

The individual string replaces are not sufficient as subtests may
introduce many other characters. Switch to a regular expression that
will replace any disallowed character with underscore.

* Move relay packet without timeout test into relayertest

* Move relay packet test with height timeout to relayertest

* Move timestamp timeout test to relayertest

* Support custom chain sets in ibctest matrix

* Prefer require.NoError

The explicit error checking was an unusual usage pattern for require.

* Add makefile with target ibctest

This puts the distributable ibctest binary in the bins directory.

* Use condensed hostnames in PeerString

Now with multiple validators or full nodes, tests don't hang.

* Try running tests with -p 1

This will limit to a single test process at a time. The tests are
failing regularly on CI despite passing locally, so I have a slight
suspicion that it has something to do with resource contention.

* Fix condenseHostName

Using ... was causing an obscure timeout in tests.

* Run tests on CI with -parallel 1

We already had -p 1 to limit to one test process at a time. Now add
-parallel 1 to only run a single parallel test at a time, to see if that
fixes the flakiness.