Add godoc badge and more package documentation (#109)
* Add godoc badge to README

* Move dockerutil package under internal directory

The dockerutil package is not intended to be part of a public API.

And add package comment.

* Add more package documentation

My rule of thumb is that a package that already has multiple files will
receive a new doc.go file containing the godoc, even if it is only a
sentence. Otherwise, code authors are left searching through multiple
files when they need to update the package documentation.

If the package only contains a single source file, I add the package
documentation in that lone file.

* Add package documentation to chain/internal/tendermint

After moving the package from chain/tendermint to
chain/internal/tendermint. As far as I can see, the tendermint package
is not intended to be used outside of the chain package. If I'm wrong
about that, it will be trivial to move the package back where it was.