Add (*Interchain).Close method (#157)
* Add Stop method to blockdb.Collector

When troubleshooting tests, the collector goroutines stay alive and
create noise for stack traces. The Stop method cleans up the goroutine
to reduce that noise.

* Add (*Interchain).Close method

This cleans up any created resources from Build(). The actual docker
containers are torn down via t.Cleanup, but the block collector
goroutine associated with the chainSet is now destroyed upon call to
Close, which reduces the number of goroutines present in a stack trace
due to a panic, test timeout, or deliberate SIGQUIT.