Switch to plain Docker SDK (#187)
* Create plain Docker client in DockerSetup

But don't yet use it anywhere.

* Switch DockerRelayer to using docker client directly

I have still seen hanging containers as of this change.

* Switch rest of dockertest use to plain Docker SDK

Unfortunately the contianer hanging issue is still present. There is one
remaining use of dockertest, whose return value is being ignored; that
use will be removed in the next commit.

* Remove remaining uses of dockertest

We still sometimes end up with stuck created containers, but I have a
lead on a ContainerCreate call that could be retried.

* Add dockerutil.StartContainer function

This attempts to gracefully handle timeouts. It doesn't address the
newly discovered underlying issue that the Docker engine becomes
unresponsive, but it seems worth keeping at least until that issue is
addressed.

* Add retry to network pruning in dockerCleanup

This seems to fix the intermittent logging about being unable to prune
networks, but it doesn't fix the hanging cotainers.

* Rename 'client' variables to 'cli'

To avoid conflicting with the client package.

* Pass context to private method